#66: 컴포넌트화 및 기존 페이지 수정
